requirements.txt
ipython-sql==0.4.0
psycopg2-binary==2.9.3
SQLAlchemy-Utils==0.38.2